Add checks for xformers and CPU only training, better status in the UI, console logging.
If xformers is installed and a cmd flag set to use it, nag the user before training that they need to enable fp16 and 8bit adam.

If user is training on CPU only, tell them they need to disable fp16 and adam.

Ensure warning messages are REALLy shown in the UI.

Better logging of installed features on launch.