d8ahazard recommendation
For the UI stuff, I think it'd make more sense to change it from "Use UNET" to "Train UNET", and put the "Train Text Encoder" setting between it and "Train in 2 stages" or whatever it's called.

Also, in the start_training method, you should add a check to the others to ensure that at least the unet or text encoder are enabled for training (or lora, idk if this affects that), but basically, a sanity check to ensure that something is going to be trained.