WORK
Add additional save settings to config.
Update saving code, not to use the new settings, but just to work better when using batches.
Move "Generate Sample Images" to dreambooth.py so we can reload it on the fly.
Hide progressbar on start of new function.
Progress bar style work/fixes.