Update UI auto-toggles, fix token padding
We don't want to try padding tokens if we're not training the text encoder, so disable that if train_tenc is False.

As we're doing this in the backend, we don't need to auto-toggle pad_tokens, so use that bit of code to instead make sure stop_tenc is 1 if not training unet.